pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13"
quart = ">=0.20.0,<0.21.0"
pydantic-settings = "^2.10.1"
quart-schema = "^0.22.0"
sqlalchemy = "^2.0.42"
aiomysql = "^0.2.0"
argon2-cffi = "^25.1.0"
pyjwt = "^2.10.1"

[tool.poetry.scripts]
app = "app.main:run"
initdb = "init_db:run"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
